Finished 4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1})
@@ -17,9 +17,8 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    if (!request.user) {
         return response.status(401).json({ error: 'Token invalid' })
     }
 
@@ -27,7 +26,7 @@ blogsRouter.post('/', async (request, response, next) => {
         return response.status(400).json({ error: 'title or url missing'})
     }
 
-    const user = await User.findById(request.body.userId)
+    const user = await User.findById(request.user)
     
     if (!user) {
         return response.status(400).json({ error: 'userId missing or not valid'})
@@ -46,9 +45,21 @@ blogsRouter.post('/', async (request, response, next) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(401).json({ error: 'blog doesnt exist'})
+    }
+    const blogUserId = blog.user.toString()
+
+    if (user && user === blogUserId) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'Cant delete blog that wasnt created by you'})
+    }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
